API Confirmação de email

+++ api/src/domain/users/application/use-cases/email-confirmation.ts
@@ -1,9 +1,11 @@
+import { Injectable } from '@nestjs/common'
 import dayjs from 'dayjs'
 
 import { Either, left, right } from '@/core/either'
 
 import { UserTempRepository } from '../repositories/user-temp-repository'
 import { CreateUserUseCase } from './create-user'
+import { AlreadyExistsEmailError } from './errors/already-exists-email-error'
 import { TokenExpiredError } from './errors/token-expired-error'
 import { TokenNotFoundError } from './errors/token-not-found-error'
 
@@ -11,8 +13,12 @@ interface EmailConfirmationUseCaseRequest {
   token: string
 }
 
-type EmailConfirmationTempUseCaseResponse = Either<TokenExpiredError, null>
+type EmailConfirmationTempUseCaseResponse = Either<
+  TokenExpiredError | TokenNotFoundError | AlreadyExistsEmailError,
+  null
+>
 
+@Injectable()
 export class EmailConfirmationUseCase {
   constructor(
     private userTempRepository: UserTempRepository,
@@ -34,12 +40,14 @@ export class EmailConfirmationUseCase {
       return left(new TokenExpiredError())
     }
 
-    if (userTemp) {
-      await this.createUserUseCase.execute({
-        email: userTemp.email,
-        name: userTemp.name,
-        password: userTemp.password,
-      })
+    const result = await this.createUserUseCase.execute({
+      email: userTemp.email,
+      name: userTemp.name,
+      password: userTemp.password,
+    })
+
+    if (result.isLeft()) {
+      return left(result.value)
     }
 
     await this.userTempRepository.delete(userTemp.id.toString())
